feat: new page for repo-rules
