+++ packages/ui/locales/en/views.json
@@ -157,15 +157,15 @@
     "rulesApplied": "rules applied",
     "bypassAllowed": "bypass allowed",
     "bypassNotAllowed": " bypass not allowed",
-    "rules": "Rules",
-    "rulesDescription": "Define standards and automate workflows to ensure better collaboration and control in your repository.",
     "newRule": "New branch rule",
     "settingsToolTip": "Cannot change settings while loading or updating.",
     "security": "Security",
     "secretScanning": "Secret scanning",
     "secretScanningDescription": "Block commits containing secrets like passwords, API keys and tokens.",
+    "rules": "Rules",
     "webhooks": "Webhooks",
     "labels": "Labels",
+    "rulesDescription": "Define standards and automate workflows to ensure better collaboration and control in your repository.",
     "cloneRepo": "Clone repository",
     "cloneHttps": "HTTPS",
     "cloneSsh": "SSH",
