+++ packages/ui/locales/fr/views.json
@@ -157,15 +157,15 @@
     "rulesApplied": "Règles appliquées",
     "bypassAllowed": "Contournement autorisé",
     "bypassNotAllowed": "Contournement non autorisé",
-    "rules": "Règles",
-    "rulesDescription": "Define standards and automate workflows to ensure better collaboration and control in your repository.",
     "newRule": "Nouvelle règle de branche",
     "settingsToolTip": "Impossible de modifier les paramètres lors du chargement ou de la mise à jour.",
     "security": "Sécurité",
     "secretScanning": "Analyse des secrets",
     "secretScanningDescription": "Empêcher les commits contenant des secrets comme les mots de passe ou jetons API.",
+    "rules": "Règles",
     "webhooks": "Webhooks",
     "labels": "Labels",
+    "rulesDescription": "Define standards and automate workflows to ensure better collaboration and control in your repository.",
     "cloneRepo": "Cloner le dépôt",
     "cloneHttps": "HTTPS",
     "cloneSsh": "SSH",
